#17 [FEAT] Read memo item list

+++ app/src/main/java/com/gimangi/singleline_note/data/database/dto/MemoTableEntity.kt
@@ -9,7 +9,7 @@ import java.util.*
 data class MemoTableEntity(
     var memoName: String,
     var suffix: String,
-    var summary: Int = 0,
+    var summary: Long = 0,
     var status: String,
     var updatedAt: Date,
     val memoList: List<MemoItemEntity>,
